requirements.txt
Flask==2.2.5
google-generativeai==0.5.2
pinecone-client==3.0.2
